HH-188298 move transactional check to another file for exclusion possibility
