+++ nab-jclient/src/main/java/ru/hh/nab/jclient/NabJClientConfig.java
@@ -8,7 +8,6 @@ import jakarta.inject.Singleton;
 import java.time.Duration;
 import java.util.Optional;
 import static java.util.Optional.ofNullable;
-import java.util.Set;
 import java.util.concurrent.ScheduledExecutorService;
 import static java.util.concurrent.TimeUnit.MINUTES;
 import ru.hh.jclient.common.HttpClientContext;
@@ -17,11 +16,10 @@ import ru.hh.jclient.common.HttpClientEventListener;
 import ru.hh.jclient.common.HttpClientFactoryBuilder;
 import ru.hh.jclient.common.check.GlobalTimeoutCheck;
 import ru.hh.jclient.common.util.storage.MDCStorage;
-import static ru.hh.nab.common.qualifier.NamedQualifier.COMMON_SCHEDULED_EXECUTOR;
 import ru.hh.nab.common.properties.FileSettings;
+import static ru.hh.nab.common.qualifier.NamedQualifier.COMMON_SCHEDULED_EXECUTOR;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
 import static ru.hh.nab.jclient.UriCompactionUtil.compactUri;
-import ru.hh.nab.jclient.checks.TransactionalCheck;
 
 @ApplicationScoped
 public class NabJClientConfig {
@@ -49,22 +47,6 @@ public class NabJClientConfig {
                 MINUTES.toMillis(sendIntervalMinutes)
             )
         ).withUserAgent(serviceName);
-  }
-
-  @Produces
-  @Singleton
-  TransactionalCheck transactionalCheck(@Named(COMMON_SCHEDULED_EXECUTOR) ScheduledExecutorService executorService, FileSettings fileSettings) {
-    var subSettings = fileSettings.getSubSettings("jclient.listener.transactional-check");
-    long sendIntervalMinutes = ofNullable(subSettings.getInteger("send.interval.minutes")).orElse(1);
-    boolean failOnCheck = ofNullable(subSettings.getBoolean("fail.on.check")).orElse(Boolean.FALSE);
-    int stacktraceDepthLimit = ofNullable(subSettings.getInteger("stacktrace.depth.limit")).orElse(10);
-    var packagesToSkip = ofNullable(subSettings.getStringList("packages.to.skip")).map(Set::copyOf).orElseGet(Set::of);
-    return new TransactionalCheck(
-        failOnCheck ? TransactionalCheck.Action.RAISE : TransactionalCheck.Action.LOG,
-        stacktraceDepthLimit,
-        executorService, MINUTES.toMillis(sendIntervalMinutes),
-        packagesToSkip
-    );
   }
 
   @Produces
